showing fetched notices
